Feature: Converted into installable module
There is a bit more work that needs to be done and a bunch of things
that changed. I left a few TODOs that will need some attention in
setup.py.

I've also added a .travis.yml, the owner of the repo will need to
setup the build on http://travis-ci.org. The travis ci build should
run all 22 unit tests as well as flake8.

The tests can be run using `python setup.py test`. You do not need to
create a virtualenv and install the requirements beforehand it will take
care of that automatically, although you can if you so choose - which I
recommend.

Also, the owner of the repo will need to register the name on
https://pypi.python.org and will need to build and upload the egg. I set
the name (and named the directory) `scaffold_vim_plugin`, the name can
be changed easily enough as well as the command which is currently
`scaffold-vim-plugin`.

All of the tests have been refactored (probably can be a bit more).

The main change in functionality is the workflow involved.

Step 1: `pip install scaffold_vim_plugin`
Step 2: execute `scaffold-vim-plugin` on the command line
Step 3: answer the prompts
Step 4: profit

The new plugin will be created in the current working directory.

...most massive commit ever...